requirements.txt
# Web Scraping & Google Trends
beautifulsoup4==4.13.3
lxml
requests==2.31.0
pytrends==4.9.2

# Apache Airflow Core
apache-airflow==2.9.0

# Airflow Providers (optional, jika digunakan dalam DAG)
apache-airflow-providers-http
apache-airflow-providers-google

# Data manipulation & date handling (opsional tapi disarankan)
pandas
python-dateutil

plotly
pandas
kaleido
wordcloud 
matplotlib 
seaborn
